extend the bootpath values to 64bits, and deal with fallout extend str2hex to handle long's, and teach it that strings starting with 'w', consisting of hex digits are really just hex digits match fp to scsibus (fiber port) match fiber channel portwwn/lun to it's bootpath variables result: with a small change to isp(mailed to mjacob) we can correctly autoconf with a boot path of: /pci@8,600000/SUNW,qlc@4/fp@0,0/disk@w21000004cf948498,0:a/bsd ie. schizo/   isp/       scsisbus/disk with port WWN of 21...98, lun 0, partition a

+++ sys/arch/sparc64/include/autoconf.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: autoconf.h,v 1.11 2006/03/04 10:31:50 miod Exp $	*/
+/*	$OpenBSD: autoconf.h,v 1.12 2006/05/31 20:11:31 jason Exp $	*/
 /*	$NetBSD: autoconf.h,v 1.10 2001/07/24 19:32:11 eeh Exp $ */
 
 /*-
@@ -147,7 +147,7 @@ struct device;
 struct bootpath {
 	char	name[16];	/* name of this node */
 	char	compatible[16];	/* "compatible" name of this node */
-	int	val[3];		/* up to three optional values */
+	long	val[3];		/* up to three optional values */
 	struct device *dev;	/* device that recognised this component */
 };
 struct bootpath	*bootpath_store(int, struct bootpath *);
